bots/requirements.txt
meshtastic>=2.0
requests>=2.28
supabase>=2.0
python-dotenv>=1.0
